Fixed CBH procedure & added saturation functionality in nightly
The hydrogenated CBH procedure broke after CBH-3. This is because the reactants were defined as the simplified products of the previous rung. The reactants should be defined as (and is now the case for the code) the unsimplified products of the previous rung NOT including residues computed from terminal atoms.
Finally, in the nightly folder, the CBH scheme has been updated to allow for different saturation atoms other than H. It currently has only been tested for halogens as the saturation atom.