CBH-0-F Handling
Automatically avoids the use of F2 as a precursor in CBH-0-F. Replaces it with an equation that utilizes CF4 and CH4. This reduces the number of total molecules needed, leading to more accurate heat of formation calculations. This only applies to molecules that have fluorine as natural substituents and might only contain C H or O. This does not apply to radicals.