Added coefficient priority for CBH schemes
CBH schemes have option to prioritize  the rungs by (sum(coeff_pdts) - sum(coeff_rcts)). This is being tested with the alternative of total number of products and reactants of a rung:
(sum(coeff_pdts) + sum(coeff_rcts))
Also corrected an error so that the minimum number of coefficients in alternative reactions were being chosen.